requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
audioread==3.0.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
deep-translator==1.11.4
distro==1.9.0
Flask==2.2.5
flask-cors==5.0.1
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==4.0.0rc1
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h2==3.2.0
h5py==3.13.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
keras==2.15.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.1.0
namex==0.0.8
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.71.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.40
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.14.1
